Minor cleanup in YAML Language Support
