+++ ide/languages.yaml/test/unit/src/org/netbeans/modules/languages/yaml/YamlTestBase.java
@@ -40,7 +40,7 @@ public class YamlTestBase extends CslTestBase {
 
     @Override
     protected String getPreferredMimeType() {
-        return YamlTokenId.YAML_MIME_TYPE;
+        return YamlLanguage.MIME_TYPE;
     }
 
     @Override
